more fixes

+++ utils/tsuite/tests/suite0/cmd/dd-small-io.sh
@@ -1,4 +1,6 @@
 #!/bin/sh
 
+[ $1 -eq 0 ] || exit 0
+
 dd if=$RANDOM bs=512 of=outfile
 diff -q $RANDOM outfile
