+++ utils/tsuite/tests/suite0/cmd/dd-weird.sh
@@ -1,5 +1,7 @@
 #!/bin/sh
 
+[ $1 -eq 0 ] || exit 0
+
 dd if=$RANDOM of=r000 seek=1M count=1 bs=64k
 md5sum r000
